+++ rent-fetch.php
@@ -3,7 +3,7 @@
 	Plugin Name: Rent Fetch
 	Plugin URI: https://github.com/jonschr/rent-fetch
     Description: Syncs properties, and floorplans with various rental APIs
-	Version: 3.9.18
+	Version: 3.10
     Author: Brindle Digital
     Author URI: https://www.brindledigital.com/
 
@@ -24,7 +24,7 @@ if ( !defined( 'ABSPATH' ) ) {
 }
 
 // Define the version of the plugin
-define ( 'RENTFETCH_VERSION', '3.9.18' );
+define ( 'RENTFETCH_VERSION', '3.10' );
 
 // Plugin directory
 define( 'RENTFETCH_DIR', plugin_dir_path( __FILE__ ) );
@@ -39,6 +39,63 @@ define( 'RENTFETCH_ACF_URL', plugin_dir_url( __FILE__ ) . 'vendor/acf/' );
 //////////////////////////////
 
 require_once( plugin_dir_path( __FILE__ ) . 'vendor/action-scheduler/action-scheduler.php' );
+
+// $table_list = array(
+//     'actionscheduler_actions',
+//     'actionscheduler_logs',
+//     'actionscheduler_groups',
+//     'actionscheduler_claims',
+// );
+
+// $found_tables = $wpdb->get_col( "SHOW TABLES LIKE '{$wpdb->prefix}actionscheduler%'" );
+
+// foreach ( $table_list as $table_name ) {
+//     if ( ! in_array( $wpdb->prefix . $table_name, $found_tables ) ) {
+//         // Table is missing, recreate it
+//         $sql = "CREATE TABLE {$wpdb->prefix}{$table_name} (
+//             // Define table structure here
+//         );";
+
+//         // Execute the SQL query to recreate the table
+//         $wpdb->query( $sql );
+
+//         // Additional actions after table recreation, if needed
+
+//         // Output a success message
+//         echo "Table {$wpdb->prefix}{$table_name} created successfully.";
+//     }
+// }
+
+
+// $table_list = array(
+//     'actionscheduler_actions',
+//     'actionscheduler_logs',
+//     'actionscheduler_groups',
+//     'actionscheduler_claims',
+// );
+
+// $found_tables = $wpdb->get_col( "SHOW TABLES LIKE '{$wpdb->prefix}actionscheduler%'" ); // phpcs:ignore WordPress.DB.PreparedSQL.InterpolatedNotPrepared
+// foreach ( $table_list as $table_name ) {
+//     if ( ! in_array( $wpdb->prefix . $table_name, $found_tables ) ) {
+        
+//         // $this->admin_notices[] = array(
+//         //     'class'   => 'error',
+//         //     'message' => __( 'It appears one or more database tables were missing. Attempting to re-create the missing table(s).' , 'action-scheduler' ),
+//         // );
+//         // add_action( 'action_scheduler/created_table', array( $store, 'set_autoincrement' ), 10, 2 );
+
+//         $store_schema  = new ActionScheduler_StoreSchema();
+//         $logger_schema = new ActionScheduler_LoggerSchema();
+//         $store_schema->register_tables( true );
+//         $logger_schema->register_tables( true );
+
+//         // remove_action( 'action_scheduler/created_table', array( $store, 'set_autoincrement' ), 10 );
+        
+//         parent::display_admin_notices();
+
+//         return;
+//     }
+// }
 
 ///////////////////
 // FILE INCLUDES //
@@ -69,8 +126,31 @@ foreach ( glob( RENTFETCH_DIR . "lib/*", GLOB_ONLYDIR ) as $dir ){
 // START THE ENGINE //
 //////////////////////
 
+function rentfetch_database_tables_missing_notice() {
+    echo '<div class="notice notice-error is-dismissible">';
+    echo '<p>' . __( 'The Action Scheduler tables appear to be missing. Please <a href="/wp-admin/tools.php?page=action-scheduler">go here</a> to regenerate those.', 'rentfetch' ) . '</p>';
+    echo '</div>';
+}
+
 add_action( 'init', 'rentfetch_start_sync' );
 function rentfetch_start_sync() {
+    
+    global $wpdb;
+
+    $table_list = array(
+        'actionscheduler_actions',
+        'actionscheduler_logs',
+        'actionscheduler_groups',
+        'actionscheduler_claims',
+    );
+        
+    $found_tables = $wpdb->get_col( "SHOW TABLES LIKE '{$wpdb->prefix}actionscheduler%'" );
+    
+    // if the tables are missing, output the notice and return
+    if ( !$found_tables ) {
+        add_action( 'admin_notices', 'rentfetch_database_tables_missing_notice' );
+        return;
+    }
     
     //* REMOVE ALL ACTIONS ADDED BY OLD APARTMENTSYNC, ADDED IN 3.0, CAN BE REMOVED IN A LATER RELEASE
     // as_unschedule_action( 'apartmentsync_do_get_yardi_property_from_api' );